ctl: Support taking credentials from standard input
Right now the VNC password and RDP username and password need
to be passed to grdctl on the command line.

That isn't ideal and it would useful to be able to pass them
directly via the terminal as well.

This commit adds some logic to do just that: read the values
from the standard input if they aren't passed in on the command line.