ctl: Support non-fixed number of arguments in subcommands
At the moment, every subcommand is expected to have a predefined,
fixed number of arguments.

We need to lighten this expectation a bit to support subcommands
that take data from either the command line or optionally the stdin.

That is necessary to support more securely passing credentials
(which will be implemented in a future commit).

This commit, modifies the argument processing code to support multiple
entries for the same subcommand, each entry with its separate allowed
number of arguments. In this way, e.g., "set-credentials" will be able
to support 0, 1 or 2 arguments.